Location inversion form (#625)
* End loc inversion

* clean

* add tests

* eudonet

* unit tests

* fix unit tests

* form design

* missing form validation group

* fixtures

* clean

* continue test integration

* continue tests

* continue tests

* continue tests

* end update measure controller test

* end unit test

* retours

* tests

* tests e2e

* increase coverage

* fix migration

* retours

* fix rebase

* Update E2E tests

* Fixes

* Show default locationNew form, ui design fixes

* Fix geocoding failure index handling

* Fix coverage

---------

+++ src/Application/Regulation/Command/DeleteMeasureCommandHandler.php
@@ -4,17 +4,26 @@ declare(strict_types=1);
 
 namespace App\Application\Regulation\Command;
 
+use App\Domain\Regulation\Exception\MeasureCannotBeDeletedException;
 use App\Domain\Regulation\Repository\MeasureRepositoryInterface;
+use App\Domain\Regulation\Specification\CanDeleteMeasures;
 
 final class DeleteMeasureCommandHandler
 {
     public function __construct(
         private MeasureRepositoryInterface $measureRepository,
+        private CanDeleteMeasures $canDeleteMeasures,
     ) {
     }
 
     public function __invoke(DeleteMeasureCommand $command): void
     {
+        $regulationOrder = $command->measure->getRegulationOrder();
+
+        if (!$this->canDeleteMeasures->isSatisfiedBy($regulationOrder->getRegulationOrderRecord())) {
+            throw new MeasureCannotBeDeletedException();
+        }
+
         $this->measureRepository->delete($command->measure);
     }
 }
